feat: add selfie components for Clean Air Forum 2026 and enhance leaderboard title styling
