+++ src/website/src/components/layout/Footer.tsx
@@ -43,7 +43,7 @@ const footerLinks = {
       label: 'Faces of Clean Air',
       href: '/africa-clean-air-forum-2026/faces-of-clean-air',
     },
-    { label: 'Selfies', href: '/africa-clean-air-forum-2026/selfies' },
+    { label: 'Clean Air Faces', href: '/africa-clean-air-forum-2026/selfies' },
     { label: 'Air Quality Quiz', href: '/africa-clean-air-forum-2026/quiz' },
     {
       label: 'Quiz Leaderboard',
