+++ src/website/src/features/clean-air-forum-2026/screens/LandingScreen.tsx
@@ -7,7 +7,6 @@ import { useRouter } from 'next/navigation';
 import { FormEvent, useState } from 'react';
 
 import AmbientBackground from '@/components/clean-air-forum-2026/AmbientBackground';
-import QrCodeButton from '@/components/clean-air-forum-2026/QrCodeButton';
 import Screen from '@/components/clean-air-forum-2026/Screen';
 import { Button } from '@/components/ui/button';
 import {
@@ -220,8 +219,6 @@ export default function LandingScreen() {
             </motion.form>
           </main>
         </div>
-
-        <QrCodeButton />
       </div>
     </Screen>
   );
