+++ src/website/src/features/clean-air-forum-2026/screens/QuizScreen.tsx
@@ -7,7 +7,6 @@ import { useRouter } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
 
 import AmbientBackground from '@/components/clean-air-forum-2026/AmbientBackground';
-import QrCodeButton from '@/components/clean-air-forum-2026/QrCodeButton';
 import Screen from '@/components/clean-air-forum-2026/Screen';
 import { Button } from '@/components/ui/button';
 import {
@@ -479,7 +478,6 @@ export default function QuizScreen() {
             </p>
           </motion.div>
         </motion.header>
-        <QrCodeButton />
       </Screen>
     );
   }
@@ -589,7 +587,6 @@ export default function QuizScreen() {
             </div>
           </div>
         </section>
-        <QrCodeButton />
       </Screen>
     );
   }
@@ -759,7 +756,6 @@ export default function QuizScreen() {
           </div>
         </div>
       </section>
-      <QrCodeButton />
     </Screen>
   );
 }
