+++ src/website/src/features/faces-of-clean-air/FacesOfCleanAirPage.tsx
@@ -1107,7 +1107,7 @@ export default function FacesOfCleanAirPage() {
               className="whitespace-nowrap text-left leading-none text-white"
             >
               {isLeaderboardStage ? (
-                <span className="text-[clamp(1.1rem,2.5vw,2.5rem)] font-extrabold tracking-[-0.045em]">
+                <span className="block truncate text-[clamp(1.1rem,2.5vw,2.5rem)] font-extrabold tracking-[-0.045em]">
                   {LEADERBOARD_TITLE}
                 </span>
               ) : (
